All code formatted, new tests added:

+++ encryption/des.c
@@ -3,8 +3,8 @@
 
 #include "des.h"
 
-#include <string.h>
 #include <openssl/evp.h>
+#include <string.h>
 
 #include "../common/types.h"
 
@@ -18,7 +18,7 @@ ccrypto_error_type des3_encrypt_with_ecb(unsigned char *key, unsigned char *data
 
     EVP_CIPHER_CTX *ctx = NULL;
     /* Create and initialise the context */
-    if(!(ctx = EVP_CIPHER_CTX_new()))
+    if (!(ctx = EVP_CIPHER_CTX_new()))
     {
         EVP_CIPHER_CTX_free(ctx);
         return CCRYPTO_ERROR_OPENSSL;
@@ -26,29 +26,29 @@ ccrypto_error_type des3_encrypt_with_ecb(unsigned char *key, unsigned char *data
 
     /* Initialise key and IV */
     /* EDE3 represents the triple des implementation */
-    if(1 != EVP_EncryptInit_ex(ctx, EVP_des_ede3_ecb(), NULL, key, NULL))
+    if (1 != EVP_EncryptInit_ex(ctx, EVP_des_ede3_ecb(), NULL, key, NULL))
     {
         EVP_CIPHER_CTX_free(ctx);
         return CCRYPTO_ERROR_OPENSSL;
     }
-    
+
     int len1 = 0;
     /* Provide the message to be encrypted, and obtain the encrypted output.
      * EVP_EncryptUpdate can be called multiple times if necessary
      */
-    if(1 != EVP_EncryptUpdate(ctx, encrypted, &len1, data, strlen(data)))
+    if (1 != EVP_EncryptUpdate(ctx, encrypted, &len1, data, (int)strlen(data)))
     {
         EVP_CIPHER_CTX_free(ctx);
         return CCRYPTO_ERROR_OPENSSL;
-    }  
+    }
 
     /* Finalise the encryption. Normally ciphertext bytes may be written at
      * this stage, but this does not occur in GCM mode
      */
     int len2 = 0;
-    if(1 != EVP_EncryptFinal_ex(ctx, encrypted + len1, &len2))
+    if (1 != EVP_EncryptFinal_ex(ctx, encrypted + len1, &len2))
     {
-        EVP_CIPHER_CTX_free(ctx);  
+        EVP_CIPHER_CTX_free(ctx);
         return CCRYPTO_ERROR_OPENSSL;
     }
 
@@ -57,7 +57,6 @@ ccrypto_error_type des3_encrypt_with_ecb(unsigned char *key, unsigned char *data
 
     return CCRYPTO_SUCCESS;
 }
-
 
 ccrypto_error_type des3_encrypt_with_cbc(unsigned char *key, unsigned char *vector, unsigned char *data, unsigned char *encrypted)
 {
@@ -69,7 +68,7 @@ ccrypto_error_type des3_encrypt_with_cbc(unsigned char *key, unsigned char *vect
 
     EVP_CIPHER_CTX *ctx = NULL;
     /* Create and initialise the context */
-    if(!(ctx = EVP_CIPHER_CTX_new()))
+    if (!(ctx = EVP_CIPHER_CTX_new()))
     {
         EVP_CIPHER_CTX_free(ctx);
         return CCRYPTO_ERROR_OPENSSL;
@@ -77,27 +76,27 @@ ccrypto_error_type des3_encrypt_with_cbc(unsigned char *key, unsigned char *vect
 
     /* Initialise key and IV */
     /* EDE3 represents the triple des implementation */
-    if(1 != EVP_EncryptInit_ex(ctx, EVP_des_ede3_cbc(), NULL, key, vector))
+    if (1 != EVP_EncryptInit_ex(ctx, EVP_des_ede3_cbc(), NULL, key, vector))
     {
         EVP_CIPHER_CTX_free(ctx);
         return CCRYPTO_ERROR_OPENSSL;
     }
-    
+
     int len1 = 0;
     /* Provide the message to be encrypted, and obtain the encrypted output.
      * EVP_EncryptUpdate can be called multiple times if necessary
      */
-    if(1 != EVP_EncryptUpdate(ctx, encrypted, &len1, data, strlen(data)))
+    if (1 != EVP_EncryptUpdate(ctx, encrypted, &len1, data, strlen(data)))
     {
         EVP_CIPHER_CTX_free(ctx);
         return CCRYPTO_ERROR_OPENSSL;
-    }  
+    }
 
     /* Finalise the encryption. Normally ciphertext bytes may be written at
      * this stage, but this does not occur in GCM mode
      */
     int len2 = 0;
-    if(1 != EVP_EncryptFinal_ex(ctx, encrypted + len1, &len2))
+    if (1 != EVP_EncryptFinal_ex(ctx, encrypted + len1, &len2))
     {
         EVP_CIPHER_CTX_free(ctx);
         return CCRYPTO_ERROR_OPENSSL;
